+++ Src/CSharpier.Rider/src/main/java/com/intellij/csharpier/CSharpierSettings.java
@@ -25,8 +25,8 @@ public class CSharpierSettings
         return this.runOnSave;
     }
 
-    public void setRunOnSave(boolean runOnSave) {
-        this.runOnSave = runOnSave;
+    public void setRunOnSave(boolean value) {
+        this.runOnSave = value;
     }
 
     private String customPath;
@@ -35,8 +35,18 @@ public class CSharpierSettings
         return this.customPath;
     }
 
-    public void setCustomPath(String customPath) {
-        this.customPath = customPath;
+    public void setCustomPath(String value) {
+        this.customPath = value;
+    }
+
+    private boolean disableCSharpierServer;
+
+    public boolean getDisableCSharpierServer() {
+        return this.disableCSharpierServer;
+    }
+
+    public void setDisableCSharpierServer(boolean value) {
+        this.disableCSharpierServer = value;
     }
 
     @Override